requirements.txt
finnhub-python
python-dotenv
google-genai
pydantic
tavily-python